+++ src/app/components/file-info-modal.js
@@ -470,55 +470,6 @@ export default Component.extend(...mixins, {
 
   fileSize: reads('file.size'),
 
-  /**
-   * One of `enabled`, `disabled`, `stopping`, `initializing`
-   * @type {ComputedProperty<String>}
-   */
-  dirStatsServiceStatus: reads('dirStatsServiceState.status'),
-
-  /**
-   * @type {ComputedProperty<boolean>}
-   */
-  isSizeStatsDisabled: computed(
-    'dirStatsServiceStatus',
-    'itemType',
-    function isSizeStatsDisabled() {
-      const {
-        dirStatsServiceStatus,
-        itemType,
-      } = this.getProperties('dirStatsServiceStatus', 'itemType');
-      return ['disabled', 'stopping'].includes(dirStatsServiceStatus) ||
-        itemType === 'symlink';
-    }
-  ),
-
-  /**
-   * @type {ComputedProperty<boolean>}
-   */
-  areSomeProvidersOffline: computed('providers', function areSomeProvidersOffline() {
-    return Boolean(this.providers?.find((p) => !p.online));
-  }),
-
-  /**
-   * @type {ComputedProperty<Array<Models.Provider>>}
-   */
-  providers: reads('providersProxy.content'),
-
-  /**
-   * List of providers that support this space.
-   * @type {ComputedProperty<Ember.Array<Provider>>}
-   */
-  providersProxy: promise.array(
-    computed('space', function providersProxy() {
-      if (this.space) {
-        return this.space.getRelation('providerList')
-          .then(providerList => get(providerList, 'list'));
-      } else {
-        resolve([]);
-      }
-    })
-  ),
-
   hardlinksCount: or('file.hardlinksCount', raw(1)),
 
   hardlinksLimitExceeded: gt('hardlinksCount', 'hardlinksLimit'),
@@ -601,22 +552,6 @@ export default Component.extend(...mixins, {
   /**
    * @type {ComputedProperty<boolean>}
    */
-  isSizeTabVisible: computed(
-    'browserModel.isDirStatsFeatureHidden',
-    'isMultiFile',
-    'file.effFile.type',
-    'itemType',
-    function isSizeTabVisible() {
-      const effItemType = get(this.file, 'effFile.type') || 'file';
-      return !this.browserModel?.isDirStatsFeatureHidden &&
-        !this.isMultiFile && effItemType !== 'file' &&
-        this.itemType !== 'symlink';
-    }
-  ),
-
-  /**
-   * @type {ComputedProperty<boolean>}
-   */
   isApiSamplesTabVisible: and(
     not('isMultiFile'),
     notEqual('itemType', raw('symlink'))
@@ -628,7 +563,6 @@ export default Component.extend(...mixins, {
   visibleTabs: computed(
     'isGeneralTabVisible',
     'isHardlinksTabVisible',
-    'isSizeTabVisible',
     'isApiSamplesTabVisible',
     'visibleTabsModels.@each.isVisible',
     function visibleTabs() {
@@ -639,9 +573,6 @@ export default Component.extend(...mixins, {
       if (this.isHardlinksTabVisible) {
         tabs.push('hardlinks');
       }
-      if (this.isSizeTabVisible) {
-        tabs.push('size');
-      }
       if (this.isApiSamplesTabVisible) {
         tabs.push('apiSamples');
       }
@@ -650,17 +581,9 @@ export default Component.extend(...mixins, {
     },
   ),
 
-  // TODO: VFS-9628 this is a temporary list of tabs moved from separate modals
-  specialFileTabs: Object.freeze([
-    'metadata',
-    'permissions',
-    'shares',
-    'qos',
-    'distribution',
-  ]),
-
   // TODO: VFS-9628 will contain all tab models after refactor
   allTabModels: collect(
+    'tabModels.size',
     'tabModels.metadata',
     'tabModels.permissions',
     'tabModels.shares',
@@ -692,7 +615,6 @@ export default Component.extend(...mixins, {
         'general',
         'apiSamples',
         'hardlinks',
-        'size',
       ];
       for (const tabId of nonModelTabIds) {
         if (!this.visibleTabs.includes(tabId)) {
@@ -723,25 +645,13 @@ export default Component.extend(...mixins, {
   }),
 
   builtInTabItems: computed(
-    'isSizeStatsDisabled',
     'hardlinksLimitExceeded',
     'hardlinksLimit',
     'hardlinksCount',
-    'areSomeProvidersOffline',
     function builtInTabItems() {
       const hardlinksCount = this.hardlinksLimitExceeded ?
         `${this.hardlinksLimit}+` :
         this.hardlinksCount;
-      let sizeStatsTabClass = '';
-      let sizeStatsStatusIcon = null;
-
-      if (this.areSomeProvidersOffline) {
-        sizeStatsTabClass = 'tab-status-warning';
-        sizeStatsStatusIcon = 'sign-warning-rounded';
-      } else if (!this.isSizeStatsDisabled) {
-        sizeStatsTabClass = 'tab-status-success';
-        sizeStatsStatusIcon = 'checkbox-filled';
-      }
 
       return {
         /** @type {FileInfoTabItem} */
@@ -758,14 +668,6 @@ export default Component.extend(...mixins, {
         },
 
         /** @type {FileInfoTabItem} */
-        size: {
-          id: 'size',
-          name: this.t('tabs.size.tabTitle'),
-          tabClass: sizeStatsTabClass,
-          statusIcon: sizeStatsStatusIcon,
-        },
-
-        /** @type {FileInfoTabItem} */
         apiSamples: {
           id: 'apiSamples',
           name: this.t('tabs.apiSamples.tabTitle'),
@@ -778,6 +680,10 @@ export default Component.extend(...mixins, {
       tabOptions: reads('fileInfoModal.tabOptions'),
       previewMode: reads('fileInfoModal.previewMode'),
       tabModelFactory: reads('fileInfoModal.tabModelFactory'),
+
+      size: computed(function size() {
+        return this.tabModelFactory.createTabModel('size');
+      }),
 
       metadata: computed(
         'tabModelFactory',
@@ -948,9 +854,6 @@ export default Component.extend(...mixins, {
     },
     toggleStorageLocations() {
       this.toggleProperty('areStorageLocationsExpanded');
-    },
-    getProvidersUrl(...args) {
-      return this.get('getProvidersUrl')(...args);
     },
   },
 });