+++ src/app/utils/filesystem-browser-model.js
@@ -43,7 +43,7 @@ export const commonActionIcons = Object.freeze({
   permissions: 'browser-permissions',
   shares: 'browser-share',
   qos: 'qos',
-  distribution: 'browser-distribution',
+  distribution: 'provider',
 });
 
 const availableButtonNames = Object.freeze([