+++ src/tests/integration/components/space-automation/atm-workflow-executions-table/workflow-row-test.js
@@ -82,7 +82,7 @@ const executionActionsSpec = [{
 }, {
   className: 'copy-record-id-action-trigger',
   label: 'Copy ID',
-  icon: 'copy',
+  icon: 'browser-copy',
 }];
 
 const scheduleTime = moment();